oauth_flow: Properly handle errors
This handles various errors during communication with the admin API,
such as connection errors and parsing errors, as well as errors returned
by the API itself (e.g. an invalid login_challenge).

We also set the ORY Hydra error endpoint to our new error page, but this
semes to only apply to some errors, e.g. if you send a GET to
/oauth2/auth, we get an invalid client error, but 404 errors are handled
by ORY Hydra directly.

Finally, errors reached within flask are also handled by this endpoint.
The main case would be when the user refuses to login at Goose. On the
other hand, this really should be caught and returned to ORY Hydra
instead.