oauth_flow: Show logged in user in consent page
Currently if one has already previously authenticated with OAuth2 and
attempts to log into a new site, they are shown the consent page
directly. If the client requests no scopes, then the user does not see
who they are logged in as.

For consistency, we show the crsid all the time.